# openapi.yaml
openapi: 3.0.3
info:
  title: ApiMe
  version: "1.3.0"
  description: |
    API para orquestração de múltiplas instâncias de WhatsApp.

    Autenticação: header `Authorization: Bearer <token>`, com três tipos de credencial
    (userJwt, apiToken, instanceToken). O middleware valida em cascata e para no primeiro
    que reconhecer o token.

    Escopo: `instanceToken` só alcança a própria instância. As rotas de perfil
    (`/instances/{id}/profile/*`, `/instances/{id}/business/*`) exigem `instanceToken` e
    recusam os demais. As rotas `/users/*` exigem usuário com role admin.
servers:
  - url: http://localhost:8080/api
    description: Local
paths:
...
  /instances/{id}/whatsapp/check:
    post:
      summary: Verificar se um número tem WhatsApp
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
      responses:
        "200":
          description: Resultado
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/presence:
    post:
      summary: Definir presença no chat
      tags: [WhatsApp Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                to:
                  type: string
                state:
                  type: string
                  description: composing, paused, available ou unavailable
      responses:
        "200":
          description: Enviada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/contacts:
    get:
      summary: Listar contatos
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Lista
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/contacts/{jid}:
    get:
      summary: Detalhar contato
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: jid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contato
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/userinfo/{jid}:
    get:
      summary: Informações públicas do usuário
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: jid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dados
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/messages/read:
    post:
      summary: Marcar mensagem como lida
      tags: [WhatsApp Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chat, message_id]
              properties:
                chat:
                  type: string
                message_id:
                  type: string
                sender:
                  type: string
                  description: Obrigatório em grupo
                played:
                  type: boolean
                  description: Marca áudio como ouvido
      responses:
        "200":
          description: Marcada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/messages/delete:
    post:
      summary: Apagar mensagem para todos
      tags: [WhatsApp Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chat, message_id]
              properties:
                chat:
                  type: string
                message_id:
                  type: string
                sender:
                  type: string
      responses:
        "200":
          description: Apagada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/messages/edit:
    post:
      summary: Editar mensagem enviada
      tags: [WhatsApp Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chat, message_id, text]
              properties:
                chat:
                  type: string
                message_id:
                  type: string
                text:
                  type: string
      responses:
        "200":
          description: Editada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/messages/react:
    post:
      summary: Reagir a uma mensagem
      tags: [WhatsApp Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chat, message_id]
              properties:
                chat:
                  type: string
                message_id:
                  type: string
                emoji:
                  type: string
                  description: Vazio remove a reação
                sender:
                  type: string
                from_me:
                  type: boolean
      responses:
        "200":
          description: Reagida
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/privacy:
    get:
      summary: Ler privacidade
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Configurações
        "403":
          description: Token não é da instância, ou não é token de instância
    post:
      summary: Alterar uma configuração de privacidade
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value]
              properties:
                name:
                  type: string
                value:
                  type: string
      responses:
        "200":
          description: Alterada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/chat-settings/{chat}:
    get:
      summary: Ler configurações do chat
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: chat
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Configurações
        "403":
          description: Token não é da instância, ou não é token de instância
    post:
      summary: Alterar configurações do chat
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: chat
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                muted_until:
                  type: string
                  description: Data ISO, ou nulo para desmutar
      responses:
        "200":
          description: Alteradas
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/status:
    post:
      summary: Definir recado do perfil
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Definido
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/disappearing-timer:
    post:
      summary: Definir timer padrão de mensagens efêmeras
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seconds]
              properties:
                seconds:
                  type: integer
                  description: 0 desliga
      responses:
        "200":
          description: Definido
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/status-privacy:
    get:
      summary: Ler privacidade do status
      tags: [WhatsApp Configurações]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Configuração
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/upload:
    post:
      summary: Subir mídia para reuso
      tags: [WhatsApp Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [media_type, data_base64]
              properties:
                media_type:
                  type: string
                  description: image, video, audio ou document
                data_base64:
                  type: string
      responses:
        "200":
          description: Enviada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/qr/contact:
    get:
      summary: Obter o link de QR do próprio contato
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: revoke
          in: query
          schema:
            type: boolean
          description: Revoga o link anterior e gera outro
      responses:
        "200":
          description: Link
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/qr/contact/resolve:
    post:
      summary: Resolver link de QR de contato
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Contato
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/qr/business-message/resolve:
    post:
      summary: Resolver link de mensagem de empresa
      tags: [WhatsApp Contatos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Destino
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups:
    get:
      summary: Listar grupos
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Lista
        "403":
          description: Token não é da instância, ou não é token de instância
    post:
      summary: Criar grupo
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                participants:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Criado
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/{group}:
    get:
      summary: Detalhar grupo
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: group
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dados
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/{group}/invite-link:
    get:
      summary: Obter link de convite
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: group
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Link
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/resolve-invite:
    post:
      summary: Ler dados de um grupo pelo link
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [link]
              properties:
                link:
                  type: string
      responses:
        "200":
          description: Dados
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/join:
    post:
      summary: Entrar em grupo por link
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [link]
              properties:
                link:
                  type: string
      responses:
        "200":
          description: Entrou
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/{group}/leave:
    post:
      summary: Sair do grupo
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: group
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saiu
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/{group}/participants:
    post:
      summary: Adicionar, remover ou promover participantes
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: group
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, participants]
              properties:
                action:
                  type: string
                  description: add, remove, promote ou demote
                participants:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Atualizados
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/groups/{group}/requests:
    get:
      summary: Listar pedidos de entrada
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: group
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista
        "403":
          description: Token não é da instância, ou não é token de instância
    post:
      summary: Aprovar ou rejeitar pedidos
      tags: [WhatsApp Grupos]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: group
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, participants]
              properties:
                action:
                  type: string
                  description: approve ou reject
                participants:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Processados
        "403":
          description: Token não é da instância, ou não é token de instância
...
  /instances/{id}/whatsapp/newsletters/{jid}/mark-viewed:
    post:
      summary: Marcar mensagens como vistas
      tags: [WhatsApp Newsletters]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: jid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_ids]
              properties:
                server_ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Marcadas
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/newsletters/{jid}/reaction:
    post:
      summary: Reagir a mensagem de newsletter
      tags: [WhatsApp Newsletters]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: jid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_id]
              properties:
                server_id:
                  type: string
                reaction:
                  type: string
                  description: Vazio remove
                message_id:
                  type: string
      responses:
        "200":
          description: Reagida
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/newsletters/{jid}/message-updates:
    post:
      summary: Buscar atualizações de mensagens
      tags: [WhatsApp Newsletters]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: jid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Atualizações
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/appstate/resync:
    post:
      summary: Ressincronizar app state
      description: >
        Descarta a cópia local das coleções informadas e baixa um snapshot novo do WhatsApp.
        Repara uma coleção que divergiu do servidor (mismatching LTHash), caso em que o whatsmeow
        para de aplicar patches e não tenta de novo sozinho. Sem corpo, ressincroniza
        critical_unblock_low (a agenda, que define quem vê um story) e regular (as listas de
        transmissão). A apime também faz isso sozinha quando uma sincronização falha, com intervalo
        mínimo entre tentativas, então esta rota é a saída manual.
      tags: [WhatsApp]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                collections:
                  type: array
                  items:
                    type: string
                    enum: [critical_block, critical_unblock_low, regular, regular_high, regular_low]
      responses:
        "200":
          description: >-
            Ressincronizado. Devolve resynced e skipped: uma coleção ressincronizada nos
            últimos 15 minutos entra em skipped em vez de sincronizar de novo, o que limita
            a frequência com que um snapshot inteiro é puxado do WhatsApp. Chamar duas vezes
            seguidas responde 200 com resynced vazio, e isso é resultado normal, não falha.
          content:
            application/json:
              schema:
                type: object
                properties:
                  resynced:
                    type: array
                    items:
                      type: string
                  skipped:
                    type: array
                    items:
                      type: string
        "400":
          description: Coleção inválida, ou instância não conectada
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/broadcast-lists:
    get:
      summary: Listar listas de transmissão
      tags: [WhatsApp Listas de Transmissão]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      description: >-
        Devolve {"lists": [...]}, e cada item traz jid, name, participants
        (pn e lid de cada destinatário), labelIds e updatedAt em RFC3339.
        A lista é local do remetente: quem recebe vê uma conversa 1:1 comum e
        nunca sabe que a lista existe. As listas vêm da sincronização de app
        state com o celular, então uma lista criada no aparelho só aparece aqui
        depois que o app state sincroniza. Não há como criar lista pela API.
      responses:
        "200":
          description: Lista
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/whatsapp/broadcast-lists/{jid}:
    get:
      summary: Detalhar lista de transmissão
      tags: [WhatsApp Listas de Transmissão]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: jid
          in: path
          required: true
          schema:
            type: string
          description: JID completo (123@broadcast) ou só o id (123)
      description: >-
        Devolve o objeto da lista direto, com jid, name, participants, labelIds
        e updatedAt. A lista só existe aqui depois da sincronização de app state
        com o celular: antes disso, a resposta é 404.
      responses:
        "200":
          description: Dados
        "403":
          description: Token não é da instância, ou não é token de instância
        "404":
          description: Lista não encontrada. Ela só aparece após a sincronização com o celular
...
components:
  securitySchemes:
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de sessão, obtido em POST /auth/login. Expira conforme JWT_EXP_HOURS.
        Identifica um usuário (claim sub).
    apiToken:
      type: http
      scheme: bearer
      description: >
        Token de integração, gerado por um admin em POST /users/{id}/token. Não expira.
        Identifica o mesmo usuário do JWT, para uso servidor a servidor.
    instanceToken:
      type: http
      scheme: bearer
      description: >
        Token com escopo de UMA instância, entregue na criação e trocável em
        POST /instances/{id}/token/rotate. Só acessa a própria instância.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Depreciado, mantido para não quebrar quem já referencia este nome.
        Use userJwt, apiToken ou instanceToken conforme o caso.
  parameters:
...
    instanceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...